pyproject.toml
[tool.poetry]
name = "bld-tasks"
version = "0.1.0"
description = ""
authors = ["Omar Desouky <omaressamdesouky2274@gmail.com>"]
readme = "README.md"
packages = [{include = "bld_tasks"}]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.1.3"
djangorestframework = "^3.14.0"
django-environ = "^0.9.0"
black = "^22.10.0"
django-extensions = "^3.2.1"
django-imagekit = "^4.1.0"
Pillow = "^9.3.0"
django-rest-knox = "^4.2.0"
pytest = "^7.2.0"
pytest-django = "^4.5.2"
django-filter = "^22.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
